+++ Avalon/Avalon.Text/IntWriteResultState.cs
@@ -55,7 +55,7 @@ public class IntWriteResultState : WriteResultState
         varBase = arg.Base;
         long varCase;
         varCase = arg.Case;
-        uint fillChar;
+        long fillChar;
         fillChar = arg.FillChar;
 
         long fillStart;
