+++ Avalon/Avalon.Text/TextWriteResultState.cs
@@ -40,7 +40,7 @@ public class TextWriteResultState : WriteResultState
 
         long varCase;
         varCase = arg.Case;
-        uint fillChar;
+        long fillChar;
         fillChar = arg.FillChar;
 
         long fillStart;
